--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -4,6 +4,7 @@
 * [CHANGE] Bump aruba, fakefs, minitest, mocha, reek, rubocop dependencies (by [@faisal][])
 * [CHANGE] Add .solargraph.yml config file to enable Solargraph LSP use in editors such as BBedit (by [@faisal][])
 * [CHANGE] Drop support for Ruby 2.7.x, 3.0.x (by [@faisal][])
+* [CHANGE] Cache the reek configuration loading for faster execution (by [@raff-s][])
 
 # v4.9.1 / 2024-04-14 [(commits)](https://github.com/whitesmith/rubycritic/compare/v4.9.0...v4.9.1)
 
@@ -454,3 +455,4 @@
 [@Fito]: https://github.com/Fito
 [@fbuys]: https://github.com/fbuys
 [@exoego]: https://github.com/exoego
+[@raff-s]: https://github.com/raff-s