--- a/lib/rubycritic/analysers/helpers/reek.rb
+++ b/lib/rubycritic/analysers/helpers/reek.rb
@@ -4,8 +4,12 @@ require 'reek'
 
 module RubyCritic
   class Reek < ::Reek::Examiner
+    def self.configuration
+      @configuration ||= ::Reek::Configuration::AppConfiguration.from_default_path
+    end
+
     def initialize(analysed_module)
-      super(analysed_module, configuration: ::Reek::Configuration::AppConfiguration.from_default_path)
+      super(analysed_module, configuration: self.class.configuration)
     end
   end
 end